+++ backend/elastic.js
@@ -8,9 +8,6 @@ import { Client } from '@elastic/elasticsearch';
 import _ from 'lodash';
 import pMap from 'p-map';
 import macros from './macros';
-import { Course } from './database/models/index';
-import hydrateSerializer from './database/serializers/hydrateSerializer';
-import Keys from '../common/Keys';
 
 const URL = macros.getEnvVariable('elasticURL') || 'http://localhost:9200';
 const client = new Client({ node: URL });
@@ -70,6 +67,7 @@ class Elastic {
    * Inserts a singular doc into an index.
    * @param  {string} indexName The index to insert into
    * @param  {Object} doc   The new doc to insert into the index.
+   * TODO: I guess this is necessary? same as below
    */
   insertDoc(indexName, doc) {
     return client.index({
@@ -82,6 +80,7 @@ class Elastic {
    * Bulk index a collection of documents using ids from hashmap
    * @param  {string} indexName The index to insert into
    * @param  {Object} map       A map of document ids to document sources to create
+   * TODO: same as below?
    */
   async bulkIndexFromMap(indexName, map) {
     const chunks = _.chunk(Object.keys(map), BULKSIZE);
@@ -98,9 +97,10 @@ class Elastic {
   }
 
   /**
-   U* Bulk update a collection of documents using ids fromhashmap
+   * Bulk update a collection of documents using ids fromhashmap
    * @param  {string} indexName The index to update into
    * @param  {Object} map       A map of document ids to document sources to update
+   * TODO: does this need the Elastic serializer? why does this exist?
    */
   async bulkUpdateFromMap(indexName, map) {
     const bulk = [];
@@ -116,6 +116,7 @@ class Elastic {
    * @param  {string} indexName Index to get from
    * @param  {string} id        ID to get
    * @return {Object} document source
+   * TODO: replace with Postgres?
    */
   async get(indexName, id) {
     return (await client.get({ index: indexName, type: '_doc', id: id })).body._source;
@@ -126,6 +127,7 @@ class Elastic {
    * @param  {string} indexName Index to get from
    * @param  {Array}  ids       Array of string ids to get
    * @return {Object} The map between doc ids and doc source
+   * TODO: replace with Postgres?
    */
   async getMapFromIDs(indexName, ids) {
     const got = await client.mget({
@@ -148,6 +150,7 @@ class Elastic {
    * @param {string} host     Host to search in
    * @param {string} subject  Subject (department) to search in
    * @param {integer} classId Class ID code to find
+   * TODO: replace with postgres
    */
   async getAllClassOccurrences(host, subject, classId) {
     const got = await client.search({
@@ -174,6 +177,7 @@ class Elastic {
    * @param {string} host     Host to search in
    * @param {string} subject  Subject (department) to search in
    * @param {integer} classId Class ID code to find
+   * TODO: replace with Postgres
    */
   async getLatestClassOccurrence(host, subject, classId) {
     const got = await client.search({
@@ -198,6 +202,7 @@ class Elastic {
 
   /*
    * Get all subjects for classes in the index
+   * TODO: replace with Postgres
    */
   async getSubjectsFromClasses() {
     const subjects = await client.search({
@@ -221,198 +226,21 @@ class Elastic {
     return _.map(subjects.body.aggregations.subjects.subjects.buckets, (subject) => { return subject.key.toLowerCase(); });
   }
 
-  /**
-   * Remove any invalid filter with the following criteria:
-   * 1. Correct key string and value type;
-   * 2. Check that { online: false } should never be in filters
-   *
-   * A sample filters JSON object has the following format:
-   * { 'NUpath': string[],
-   *   'college': string[],
-   *   'subject': string[],
-   *   'online': boolean,
-   *   'classType': string }
-   *
-   * @param {object} filters The json object represting all filters on classes
-   */
-  validateFilters(filters) {
-    const isString = (givenVar) => {
-      return typeof givenVar === 'string';
-    };
-
-    const isStringArray = (givenVar) => {
-      return Array.isArray(givenVar) && givenVar.every((eachVar) => isString(eachVar));
-    };
-
-    const isTrue = (givenVar) => {
-      return typeof givenVar === 'boolean' && givenVar;
-    };
-
-    const validFiltersFormat = {
-      NUpath: isStringArray,
-      college: isStringArray,
-      subject: isStringArray,
-      online: isTrue,
-      classType: isString,
-    };
-
-    const validFilters = {};
-    for (const [filterKey, filterValues] of Object.entries(filters)) {
-      if (!(filterKey in validFiltersFormat)) {
-        macros.log('Invalid filter key.', filterKey);
-      } else if (!(validFiltersFormat[filterKey](filterValues))) {
-        macros.log('Invalid filter value type.', filterKey);
-      } else {
-        validFilters[filterKey] = filterValues;
-      }
-    }
-    return validFilters;
-  }
 
   /**
-   * Get elasticsearch query from json filters and termId
-   * @param  {string}  termId  The termId to look within
-   * @param  {object}  filters The json object representing all filters on classes
+   * Return a set of all available subjects
+   * TODO: replace with Postgres
    */
-  getClassFilterQuery(termId, filters) {
-    const hasSectionsFilter = { exists: { field: 'sections' } };
-
-    const termFilter = { term: { 'class.termId': termId } };
-
-    const getNUpathFilter = (selectedNUpaths) => {
-      const NUpathFilters = selectedNUpaths.map((eachNUpath) => ({ match_phrase: { 'class.classAttributes': eachNUpath } }));
-      return { bool: { should: NUpathFilters } };
-    };
-
-    const getCollegeFilter = (selectedColleges) => {
-      const collegeFilters = selectedColleges.map((eachCollege) => ({ match_phrase: { 'class.classAttributes': eachCollege } }));
-      return { bool: { should: collegeFilters } };
-    };
-
-    const getSubjectFilter = (selectedSubjects) => {
-      const subjectFilters = selectedSubjects.map((eachSubject) => ({ match: { 'class.subject': eachSubject } }));
-      return { bool: { should: subjectFilters } };
-    };
-
-    // note that { online: false } is never in filters
-    const getOnlineFilter = (selectedOnlineOption) => {
-      return { term: { 'sections.online': selectedOnlineOption } };
-    };
-
-    const getClassTypeFilter = (selectedClassType) => {
-      return { match: { 'class.scheduleType': selectedClassType } };
-    };
-
-    // construct compound class filters (fliters joined by AND (must), options for a filter joined by OR (should))
-    const filterToEsQuery = {
-      NUpath: getNUpathFilter,
-      college: getCollegeFilter,
-      subject: getSubjectFilter,
-      online: getOnlineFilter,
-      classType: getClassTypeFilter,
-    };
-
-    const classFilters = [hasSectionsFilter, termFilter];
-    for (const [filterKey, filterValues] of Object.entries(filters)) {
-      if (filterKey in filterToEsQuery) {
-        classFilters.push(filterToEsQuery[filterKey](filterValues));
-      }
-    }
-
-    return { bool:{ must: classFilters } };
-  }
-
-  /**
-   * Search for classes and employees
-   * @param  {string}  query   The search to query for
-   * @param  {string}  termId  The termId to look within
-   * @param  {integer} min     The index of first document to retreive
-   * @param  {integer} max     The index of last document to retreive
-   * @param  {object}  filters The json object representing all filters on classes
-   */
-  async search(query, termId, min, max, filters = {}) {
+  async getSubjects() {
     if (!this.subjects) {
       this.subjects = new Set(await this.getSubjectsFromClasses());
     }
+    return this.subjects;
+  }
 
-    // if we know that the query is of the format of a course code, we want to do a very targeted query against subject and classId: otherwise, do a regular query.
-    const courseCodePattern = /^\s*([a-zA-Z]{2,4})\s*(\d{4})?\s*$/i;
-    let fields = [
-      'class.name^2', // Boost by 2
-      'class.name.autocomplete',
-      'class.subject^4',
-      'class.classId^3',
-      'sections.profs',
-      'class.crns',
-      'employee.name^2',
-      'employee.emails',
-      'employee.phone',
-    ];
-
-    const patternResults = query.match(courseCodePattern);
-    if (patternResults && this.subjects.has(patternResults[1].toLowerCase())) {
-      // after the first result, all of the following results should be of the same subject, e.g. it's weird to get ENGL2500 as the second or third result for CS2500
-      fields = ['class.subject^10', 'class.classId'];
-    }
-
-    // get compound class filters
-    const validFilters = this.validateFilters(filters);
-    const classFilters = this.getClassFilterQuery(termId, validFilters);
-
-    // text query from the main search box
-    const matchTextQuery = {
-      multi_match: {
-        query: query,
-        type: 'most_fields', // More fields match => higher score
-        fuzziness: 'AUTO',
-        fields: fields,
-      },
-    };
-
-    // use lower classId has tiebreaker after relevance
-    const sortByClassId = { 'class.classId.keyword': { order: 'asc', unmapped_type: 'keyword' } };
-
-    // filter by type employee
-    const isEmployee = { term: { type: 'employee' } };
-
-    // compound query for text query and filters
-    const mainQuery = {
-      index: `${this.EMPLOYEE_INDEX},${this.CLASS_INDEX}`,
-      from: min,
-      size: max - min,
-      body: {
-        sort: ['_score', sortByClassId],
-        query: {
-          bool: {
-            must: matchTextQuery,
-            filter: {
-              bool: {
-                should: [
-                  classFilters,
-                  isEmployee,
-                ],
-              },
-            },
-          },
-        },
-      },
-    };
-
-    const searchOutput = await client.search(mainQuery);
-
-    const resultModels = await Course.findAll({ where: { id: searchOutput.body.hits.hits.map((hit) => { return hit._id; }) } });
-    const resultScores = searchOutput.body.hits.hits.reduce((acc, elem) => {
-      acc[elem._id] = elem._score;
-      return acc;
-    }, {});
-
-    const results = await hydrateSerializer.bulkSerialize(resultModels);
-
-    return {
-      searchContent: Object.values(results).map((res) => { return { ...res, score: resultScores[Keys.getClassHash(res.class)] }; }),
-      resultCount: searchOutput.body.hits.total.value,
-      took: searchOutput.body.took,
-    };
+  async query(index, from, size, body) {
+    macros.log(JSON.stringify({ index, from, size, body }));
+    return client.search({ index, from, size, body });
   }
 }
